bug/kgpreview/3.0.0/xiezhenxiang

See merge request kgcloud/kg-cloud-kgms!106

+++ kgpreview/src/main/java/com/plantdata/kgcloud/domain/j2r/entity/Setting.java
@@ -13,7 +13,7 @@ import java.util.List;
 @ApiModel("配置参数")
 public class Setting {
 
-    public Integer dataSetId;
+    public Long dataSetId;
     public String userId;
     public String kgName;
     public String name;
